+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [*Unreleased*]
+
+### Changes
+
+* Updated to bot framework `^4.9.0`
+
 ## [*1.6.2*] - <*2020-05-17*>
 
 ### Changes